Add start of /p/{pageId} handler
This will allow get and post, eventually, but right now it
simply does a get of some simple html.

The gabbi harness is only sort of working. Would probably work
better extrated to a shell script, but for now I'm trying to
avoid that.

Next step is figuring out templates. Including if it is
possible to store them in the compiled binary easily. Maybe
something like https://github.com/gobuffalo/packr