Add very simple wiki word parsing
This leaves formatting as pre, but allows simple wikiwords to
be turned into anchors.

This leaves a big fixme: any HTML in the input will be preserved
so it is possible for the wikitext to be dangerous. It needs to
be bleached before being written to disk or otherwise manipulated.