Add tests (and fixes) for utf8 page handling
This led to the discovery that the wikiword regexp was not working
as expected for unicode page names. A unit-like test was added for
that and a corresponding Makefile target.

Once that was fixed it was then discovered that though IRIs can
work the html template system will percent encode things in a
context sensitive way, so we need to be sure to decode that
encoding when getPage and setPage are called.