Page rw locking
This uses a map of sync.RWMutex, keyed by page id, with the entire map
guarded by its own Mutex. This means that a page file can be read by
multiple readers, but not while it is being written and only one writer
may write at a time.

This is a messy solution that will need to change if/when revisions or
git-like storage happens: the lock prevents concurrent writes, but does
nothing to prevent the lost update problem.

The reason for doing it at all, knowing that it will likely go away,
was to get a better sense of how named mutexes could be made to work.
The test added to page_test gives an example of concurrency and
blocking.

There are things about this that feel messy:

* The page map is global and so is its lock.
* Something feels wrong and heavy handed about the nil handling in
  page to mutex map. I though the nil value would be an unlocked
  mutex, but it is not, it is nil, thus the dance to make sure one
  gets assigned.
* Lots of duplicated code.
* Plenty of other stuff that I'm sure to be missing.

Fixes #1