Initial stab at websocket driven recent changes stream
This almost works, but for one rather dramatic flaw:

There is a single channel of change notifications. When
one web socket client consumes a message off that channel,
that's the only client that gets it: becuase they took it.

So there needs to be a way to fanout, which can probably
come from looking at one of the other gorilla examples

but too tired now