Correct recent changes websocket stream
The fix was to use a Hub, as borrowed from a gorrilla/websocket
example [1]. This keeps track of all open connections and re-
broadcasts messages to each one (on their own goroutine).

Each change is sent to the hub as message to be broadcast, that
message is duplicated down each client's send channel.

[1] https://github.com/gorilla/websocket/tree/master/examples/chat